Limit Except and Max_client_body_size incorporated

+++ inc/ServerConfig.hpp
@@ -6,7 +6,7 @@
 /*   By: gleal <gleal@student.42lisboa.com>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/22 00:50:11 by fmeira            #+#    #+#             */
-/*   Updated: 2022/09/05 23:13:56 by gleal            ###   ########.fr       */
+/*   Updated: 2022/09/06 16:30:59 by gleal            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -87,7 +87,7 @@ class BaseConfig{
     const std::string&              get_root( void ) const;
     AutoBool                        get_autoindex( void ) const;
     const ErrorPage&                get_error_pages( void ) const;
-    int                             get_max_body_size( void ) const;
+    long long						get_max_body_size( void ) const;
     const std::vector<std::string>& get_indexes( void ) const;
     const std::vector<Redirect>&    get_redirect( void );
 
@@ -95,7 +95,7 @@ class BaseConfig{
         std::string                 _root;
         AutoBool                    _autoindex;
         ErrorPage                   _error_pages;
-        int                         _client_max_body_size;
+        long long                   _client_max_body_size;
         std::vector<std::string>    _indexes;
         std::vector<Redirect>       _redirect;
 };
